move "getSelectedDevice" from reducersUtils to deviceUtils
